backend/pyproject.toml
[project]
name = "dsn-chatbot-backend"
version = "0.1.0"
description = "DSN Chatbot – FastAPI backend"
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    # Web framework
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    # Database
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.15,<2",
    # Auth
    "python-jose[cryptography]>=3.4,<4",
    "passlib[bcrypt]>=1.7,<2",
    "python-multipart>=0.0.20,<1",
    # LLM
    "openai>=1.76,<2",
    # Settings
    "pydantic-settings>=2.9,<3",
    # Validation
    "email-validator>=2.2,<3",
    # Utilities
    "python-dotenv>=1.1,<2",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio",
    "httpx",                # TestClient async support
    "coverage[toml]",
    "pre-commit",
    "pip-tools",            # pip-compile to regenerate requirements.txt
]

# ---------------------------------------------------------------------------
# Build system
# ---------------------------------------------------------------------------
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ---------------------------------------------------------------------------
# Tool configuration
# ---------------------------------------------------------------------------

# ---- Ruff (linter + formatter) --------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "T20", # flake8-print
]
ignore = [
    "E501", # line-length handled by formatter
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ---- Mypy -----------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true

# ---- Pytest ---------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

# ---- Coverage -------------------------------------------------------------
[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 80
